refactor: 收敛工具路由边界
