+++ qq-maid-core/src/runtime/respond/mod.rs
@@ -56,6 +56,7 @@ mod title;
 mod tool_presenters;
 mod tool_projection;
 mod tool_route;
+mod tool_route_domains;
 mod tool_runtime;
 mod train_flow;
 mod translation_flow;
